CNN_Image_classifier/app.py
import streamlit as st
import numpy as np
import tensorflow as tf
from PIL import Image

# Load the trained model
model = tf.keras.models.load_model("cnn_model.keras")

# CIFAR-10 class labels
class_names = ['airplane', 'automobile', 'bird', 'cat', 'deer',
               'dog', 'frog', 'horse', 'ship', 'truck']

st.title("Image Classifier: CIFAR-10 Categories")
st.write("Upload an image and let the CNN classify it.")

# Upload image
uploaded_file = st.file_uploader("Choose an image...", type=["jpg", "jpeg", "png"])

if uploaded_file is not None:
    # Show the uploaded image
    image = Image.open(uploaded_file).convert("RGB")
    st.image(image, caption="Uploaded Image", use_column_width=True)

    # Preprocess the image
    img = image.resize((32, 32))
    img_array = np.array(img) / 255.0  # normalize
    img_array = img_array.reshape((1, 32, 32, 3))  # model expects 4D input

    # Make prediction
    prediction = model.predict(img_array)
    predicted_class = class_names[np.argmax(prediction)]
    confidence = np.max(prediction)

    st.subheader(f"Prediction: **{predicted_class}**")
    st.write(f"Confidence: {confidence * 100:.2f}%")